Add chunk-based inference
