--- config/enhancer_stage1.yaml
@@ -1,4 +1,0 @@
-lcfm_training_mode: ae
-load_fg_only: true
-batch_size_per_gpu: 16
-denoiser_run_dir: runs/denoiser
