--- config/enhancer_stage2.yaml
@@ -1,8 +1,0 @@
-lcfm_training_mode: cfm
-batch_size_per_gpu: 32
-training_seconds: 3.0
-gan_training_start_step: null
-lcfm_z_scale: 6
-praat_augment_prob: 0.2
-denoiser_run_dir: runs/denoiser
-enhancer_stage1_run_dir: runs/enhancer_stage1
