--- packages.txt
@@ -1,1 +1,0 @@
-libsox-dev
