--- pyproject.toml
@@ -1,6 +1,0 @@
-[tool.black]
-line-length = 120
-target-version = ['py310']
-
-[tool.isort]
-line_length = 120
