+++ requirements.txt
@@ -1,18 +1,3 @@
-celluloid>=0.2.0
-deepspeed>=0.12.4
-librosa>=0.10.1
-matplotlib>=3.8.1
-numpy>=1.26.2
-omegaconf>=2.3.0
-pandas>=2.1.3
-ptflops>=0.7.1.2
-rich>=13.7.0
-scipy>=1.11.4
-soundfile>=0.12.1
-torch>=2.1.1
-torchaudio>=2.1.1
-torchvision>=0.16.1
-tqdm>=4.66.1
-resampy>=0.4.2
-tabulate>=0.8.10
-gradio>=4.8.0
+torch==2.0.0
+torchvision==0.15.1
+torchaudio==2.0.1
