--- resemble_enhance/data/distorter/__init__.py
@@ -1,1 +1,0 @@
-from .distorter import Distorter
