--- resemble_enhance/enhancer/lcfm/__init__.py
@@ -1,2 +1,0 @@
-from .irmae import IRMAE
-from .lcfm import CFM, LCFM
