--- resemble_enhance/enhancer/univnet/__init__.py
@@ -1,1 +1,0 @@
-from .univnet import UnivNet
